reworked endpoints to different files code is now more readable
