+++ api/.env.example
@@ -4,10 +4,17 @@
 # For real use, make sure you've completed the steps required to set up MariaDB and
 # imported the example utasuki database. The .sql file can be found in /resources/
 #
-# When done, rename this file to ".env". This file is later sourced in the start scripts
+# When done, rename this file to ".env".
 
 export DB_USER="utasuki"
 export DB_PASS="<YOUR_PASSWORD>"
 export DB_NAME="utasuki_example"
 # used for auth token generation. see https://jwt.io/
 export TOKEN_SECRET=""
+
+export DEV_PORT=5900 # ごくっ
+# paths to HTTPS/TLS/SSL certificates
+# for localhost, I use mkcert, see https://github.com/FiloSottile/mkcert
+# these should already exist if you followed the steps correctly
+export DEV_PRIVATE_KEY="cert/localhost-key.pem"
+export DEV_CERTIFICATE="cert/localhost.pem"
