+++ api/.gitignore
@@ -4,7 +4,6 @@ node_modules
 .env
 .env.*
 !.env.example
-start_prod.sh
 tmux_start_prod.sh
 cert/*
 !cert/.gitkeep
