+++ api/package.json
@@ -1,17 +1,18 @@
 {
   "name": "utasuki-api",
-  "version": "0.1.2-beta",
+  "version": "0.1.3-beta",
   "description": "REST API for UtaSuki",
   "main": "api.js",
   "scripts": {
-    "dev": "./start_dev.sh",
-    "test:dev": "npx jest",
-    "deploy:prod": "./start_prod.sh"
+    "dev": "APP_ENV=dev nodemon api.js",
+    "test:dev": "APP_ENV=dev npx jest",
+    "deploy:prod": "APP_ENV=prod node api.js"
   },
   "author": "Toralv",
   "license": "ISC",
   "dependencies": {
     "bcryptjs": "^3.0.2",
+    "dotenv": "^16.5.0",
     "express": "^4.19.2",
     "express-validator": "^7.1.0",
     "jsonwebtoken": "^9.0.2",
