+++ api/routes/tracks.js
@@ -1,0 +1,85 @@
+"use strict"
+
+const sendStatus = require("../helpers.js").sendStatus;
+const dbQuery = require("../db.js").dbQuery;
+
+const express = require("express");
+const app = express();
+
+module.exports = app.get('/', async (req, res) => {
+	let year = "";
+	let uid = 0;
+	try {
+		year = JSON.parse(req.query.data).year;
+		uid = JSON.parse(req.query.data).uid;
+	} catch (e) {
+		sendStatus(req, res, 400, "error.year_user_not_specified");
+		return;
+	}
+
+	let userExist = await dbQuery("SELECT public FROM users WHERE uid = ?", [uid]);
+	if (!userExist.length) {
+		sendStatus(req, res, 404, "error.user_not_exist");
+		return;
+	}
+
+	if (!userExist[0].public && uid != req.profile?.uid) {
+		sendStatus(req, res, 403, "error.user_private");
+		return;
+	}
+	
+	let data = await dbQuery(`
+							SELECT
+								date,
+								artist,
+								album,
+								title,
+								released,
+								tracks.image,
+								notes,
+								last_edit,
+								track_notes_public
+							FROM
+								user_tracks
+							JOIN
+								tracks
+							ON
+								tracks.id = track_id
+							JOIN
+								users
+							ON
+								users.uid = user_tracks.uid
+							WHERE users.uid = ?
+								AND date >= ?
+									AND date < ?;`,
+		[
+			uid,
+			`${year}-01-01`,
+			`${Number(year) + 1}-01-01`,
+		]);
+
+	if (data.length == 0) {
+		sendStatus(req, res, 404, "warning.no_tracks_year");
+		return;
+	}
+
+	// I'll leave this here for memories sake
+	//let monthTracks = [[], [], [], [], [], [], [], [], [], [], [], []];
+	//let monthTracks = new Array(12).fill([]);
+	let monthTracks = new Array(12).fill(null).map(() => []);
+
+	for (let i = 0; i < data.length; i++) {
+		monthTracks[data[i].date.getMonth()].push({
+			artist: data[i].artist,
+			album: data[i].album,
+			title: data[i].title,
+			released: data[i].released,
+			image: data[i].image,
+			notes: uid == req.profile?.uid ? data[i].notes : data[i].track_notes_public && data[i].notes,
+			last_edit: data[i].track_notes_public && data[i].last_edit,
+		});
+	}
+
+	sendStatus(req, res, 200, "success.tracks_found", monthTracks);
+});
+
