+++ website/package.json
@@ -6,7 +6,7 @@
 		"dev": "APP_ENV=dev NODE_TLS_REJECT_UNAUTHORIZED=0 vite dev --host",
 		"build": "vite build",
 		"preview": "vite preview",
-		"test": "vitest",
+		"test": "APP_ENV=dev vitest",
 		"deploy:prod": "vite build && vite preview --host"
 	},
 	"devDependencies": {
