+++ website/src/lib/globals.js
@@ -1,7 +1,7 @@
 import { dev } from "$app/environment";
 
 export const API_PORT = dev ? 5900 : 8800;
-export const CDN_ADDR = dev ? "https://localhost:5900" : "https://cdn.utasuki.toralv.dev";
+export const CDN_ADDR = dev ? "https://localhost:5900" : "https://cdn.utasuki.com";
 export const LEN_LIMITS = {
 	ALBUM: 255 - 5, // album name is used for the album cover's image. ext4 maximum filename length minus possible file extensions (including '.')
 	GENERAL: 255,
